+++ pkg/processor/runtime/rpc/connection/abstract.go
@@ -37,7 +37,7 @@ import (
 	"github.com/nuclio/nuclio/pkg/processor/runtime/rpc/controlmessagebroker"
 	"github.com/nuclio/nuclio/pkg/processor/runtime/rpc/encoder"
 	"github.com/nuclio/nuclio/pkg/processor/runtime/rpc/result"
-	"github.com/nuclio/nuclio/pkg/processor/statistics/metrics"
+	"github.com/nuclio/nuclio/pkg/processor/statistics"
 
 	"github.com/mitchellh/mapstructure"
 	"github.com/nuclio/errors"
@@ -89,7 +89,7 @@ func (bc *AbstractConnectionManager) UpdateStatistics(durationSec float64) {
 	bc.Configuration.Statistics.DurationMilliSecondsSum += uint64(durationSec * 1000)
 }
 
-func (bc *AbstractConnectionManager) GetAllocationStatistics() *metrics.AllocatorStatistics {
+func (bc *AbstractConnectionManager) GetAllocationStatistics() *statistics.AllocatorStatistics {
 	return bc.allocator.GetStatistics()
 }
 
@@ -473,13 +473,13 @@ func (be *AbstractEventConnection) RunHandler() {
 }
 
 // GetStatistics returns a pointer to the statistics object. This must not be modified by the reader
-func (be *AbstractEventConnection) GetStatistics() *metrics.EventProcessingStatistics {
+func (be *AbstractEventConnection) GetStatistics() *statistics.EventProcessingStatistics {
 	return nil
 }
 
 // GetAllocationStatistics returns the statistics of the allocator if there is any in the runtime
 // AbstractEventConnection runtime doesn't have any allocator in it, so return nil
-func (be *AbstractEventConnection) GetAllocationStatistics() *metrics.AllocatorStatistics {
+func (be *AbstractEventConnection) GetAllocationStatistics() *statistics.AllocatorStatistics {
 	return nil
 }
