+++ pkg/processor/runtime/rpc/connection/types.go
@@ -26,7 +26,7 @@ import (
 	"github.com/nuclio/nuclio/pkg/processor/runtime"
 	"github.com/nuclio/nuclio/pkg/processor/runtime/rpc/encoder"
 	"github.com/nuclio/nuclio/pkg/processor/runtime/rpc/result"
-	"github.com/nuclio/nuclio/pkg/processor/statistics/metrics"
+	"github.com/nuclio/nuclio/pkg/processor/statistics"
 
 	"github.com/nuclio/logger"
 )
@@ -56,7 +56,7 @@ type ConnectionManager interface {
 	UpdateStatistics(durationSec float64)
 
 	// GetAllocationStatistics retrieves the current statistics of the ConnectionManager allocator
-	GetAllocationStatistics() *metrics.AllocatorStatistics
+	GetAllocationStatistics() *statistics.AllocatorStatistics
 
 	// SetStatus updates the operational status of the ConnectionManager
 	SetStatus(status.Status)
