+++ pkg/processor/runtime/runtime.go
@@ -23,7 +23,7 @@ import (
 	"github.com/nuclio/nuclio/pkg/common/status"
 	"github.com/nuclio/nuclio/pkg/processor/controlcommunication"
 	"github.com/nuclio/nuclio/pkg/processor/databinding"
-	"github.com/nuclio/nuclio/pkg/processor/statistics/metrics"
+	"github.com/nuclio/nuclio/pkg/processor/statistics"
 
 	"github.com/nuclio/errors"
 	"github.com/nuclio/logger"
@@ -47,7 +47,7 @@ type Runtime interface {
 
 	// GetAllocationStatistics returns statistics gathered by the allocator
 	// if there is any in the runtime, otherwise returns nil
-	GetAllocationStatistics() *metrics.AllocatorStatistics
+	GetAllocationStatistics() *statistics.AllocatorStatistics
 
 	// GetConfiguration returns the runtime configuration
 	GetConfiguration() *Configuration
