+++ pkg/processor/runtime/shell/runtime.go
@@ -31,7 +31,7 @@ import (
 	"github.com/nuclio/nuclio/pkg/common/status"
 	"github.com/nuclio/nuclio/pkg/functionconfig"
 	"github.com/nuclio/nuclio/pkg/processor/runtime"
-	"github.com/nuclio/nuclio/pkg/processor/statistics/metrics"
+	"github.com/nuclio/nuclio/pkg/processor/statistics"
 
 	"github.com/nuclio/errors"
 	"github.com/nuclio/logger"
@@ -179,7 +179,7 @@ func (s *shell) ProcessBatch(batch []nuclio.Event, functionLogger logger.Logger)
 
 // GetAllocationStatistics returns the statistics of the allocator if there is any in the runtime
 // shell runtime doesn't have any allocator in it, so return nil
-func (s *shell) GetAllocationStatistics() *metrics.AllocatorStatistics {
+func (s *shell) GetAllocationStatistics() *statistics.AllocatorStatistics {
 	return nil
 }
