+++ pkg/processor/statistics/types.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package metrics
+package statistics
 
 type EventProcessingStatistics struct {
 	EventsHandledSuccess uint64
