+++ pkg/processor/trigger/types.go
@@ -24,7 +24,7 @@ import (
 
 	"github.com/nuclio/nuclio/pkg/functionconfig"
 	"github.com/nuclio/nuclio/pkg/processor/runtime"
-	"github.com/nuclio/nuclio/pkg/processor/statistics/metrics"
+	"github.com/nuclio/nuclio/pkg/processor/statistics"
 	"github.com/nuclio/nuclio/pkg/processor/util/partitionworker"
 
 	"github.com/nuclio/errors"
@@ -183,7 +183,7 @@ func (c *Configuration) ResolveWorkerAllocationMode(modeFromAttributes, modeFrom
 type Statistics struct {
 	EventsHandledSuccessTotal uint64
 	EventsHandledFailureTotal uint64
-	WorkerAllocatorStatistics metrics.AllocatorStatistics
+	WorkerAllocatorStatistics statistics.AllocatorStatistics
 }
 
 func (s *Statistics) DiffFrom(prev *Statistics) Statistics {
